chore(cmd): Move FlagClientAgentPort and flagTargetDefaultPort to uint16

+++ internal/cmd/base/servers.go
@@ -122,7 +122,7 @@ type Server struct {
 	DevSecondaryTargetId             string // Target using host sources.
 	DevHostAddress                   string // Host address for target using host sources.
 	DevTargetAddress                 string // Network address for target with address.
-	DevTargetDefaultPort             int64
+	DevTargetDefaultPort             uint16
 	DevTargetSessionMaxSeconds       int64
 	DevTargetSessionConnectionLimit  int64
 	DevLoopbackPluginId              string
